Move CC Scheme dataset to separate file.
