+++ src/sec_certs/dataset/__init__.py
@@ -1,6 +1,7 @@
 """This package exposes Datasets of various Samples, both primary (Common Criteria, FIPS) and auxiliary (CVEs, CPEs, ...)"""
 
-from sec_certs.dataset.cc import CCDataset, CCDatasetMaintenanceUpdates, CCSchemeDataset
+from sec_certs.dataset.cc import CCDataset, CCDatasetMaintenanceUpdates
+from sec_certs.dataset.cc_scheme import CCSchemeDataset
 from sec_certs.dataset.cpe import CPEDataset
 from sec_certs.dataset.cve import CVEDataset
 from sec_certs.dataset.fips import FIPSDataset
